refactor and fix one more threadsafety issue
