+++ tools/cli/domain.go
@@ -83,14 +83,13 @@ func newDomainCommands() []cli.Command {
 				newDomainCLI(c, false).DescribeDomain(c)
 			},
 		},
-
 		{
 			Name:    "migration",
 			Aliases: []string{"mi"},
 			Usage:   "Migrate existing domain to new domain. This command only validates the settings. It does not perform actual data migration",
 			Flags:   migrateDomainFlags,
 			Action: func(c *cli.Context) {
-				newDomainCLI(c, true).MigrateDomain(c)
+				newDomainMigrationCLIImpl(c).Validation(c)
 			},
 		},
 	}
